+++ nanobot/nanobot.py
@@ -6,7 +6,7 @@ from dataclasses import dataclass
 from pathlib import Path
 from typing import Any
 
-from nanobot.agent.hook import AgentHook, AgentHookContext
+from nanobot.agent.hook import AgentHook, SDKCaptureHook
 from nanobot.agent.loop import AgentLoop
 from nanobot.bus.queue import MessageBus
 
@@ -104,7 +104,7 @@ class Nanobot:
                 Different keys get independent history.
             hooks: Optional lifecycle hooks for this run.
         """
-        capture = _SDKCaptureHook()
+        capture = SDKCaptureHook()
         prev = self._loop._extra_hooks
         base_hooks = list(hooks) if hooks is not None else list(prev or [])
         self._loop._extra_hooks = [capture, *base_hooks]
@@ -121,25 +121,6 @@ class Nanobot:
             tools_used=capture.tools_used,
             messages=capture.messages,
         )
-
-
-class _SDKCaptureHook(AgentHook):
-    """Record tool names and the final message list for ``RunResult``.
-
-    The runner mutates ``context.messages`` in place across iterations, so the
-    snapshot is refreshed on every ``after_iteration`` call; the last call
-    reflects the end-of-turn state the SDK caller cares about.
-    """
-
-    def __init__(self) -> None:
-        super().__init__()
-        self.tools_used: list[str] = []
-        self.messages: list[dict[str, Any]] = []
-
-    async def after_iteration(self, context: AgentHookContext) -> None:
-        for call in context.tool_calls:
-            self.tools_used.append(call.name)
-        self.messages = list(context.messages)
 
 
 def _make_provider(config: Any) -> Any:
